functions/monitoring/realtime_completeness_checker/requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
